Remove redundant cast

+++ src/fann_train.c
@@ -789,7 +789,7 @@ void fann_update_weights_sarprop(struct fann *ann, unsigned int epoch, unsigned 
 	float step_error_shift = ann->sarprop_step_error_shift; /* ld 3 = 1.585 */
 	float T = ann->sarprop_temperature;
 	float MSE = fann_get_MSE(ann);
-	float RMSE = (float)sqrt(MSE);
+	float RMSE = sqrtf(MSE);
 
 	unsigned int i = first_weight;
